feat: implement global rate limiting for agent and REST APIs
- Added AGENT_RATE_LIMIT environment variable to configure rate limits for the agent API.
- Integrated rate limiting middleware in both agent and REST routes to handle excessive requests, returning a 429 status when limits are exceeded.
- Updated example environment file to reflect new rate limit configuration.
