+++ docker-compose.yaml
@@ -28,6 +28,7 @@ services:
       # AGENT API
       - AGENT_ENABLED=true
       - AGENT_PREFIX=/agent
+      - AGENT_RATE_LIMIT=10
       - AGENT_BLACKLISTED_TABLES=
 
       # MCP Server
