+++ server/.env.example
@@ -19,6 +19,7 @@ REST_BLACKLISTED_TABLES= # ex: users,logs
 # AGENT API
 AGENT_ENABLED=true
 AGENT_PREFIX=/agent
+AGENT_RATE_LIMIT=10
 AGENT_BLACKLISTED_TABLES= # Ex: users,logs
 
 # MCP Server
